Mod download failure

+++ backend/app/utils/helpers.py
@@ -204,16 +204,22 @@ class Arma3ModManager:
         :param dst_dir: - STR, the destination directory to move the downloaded mod to
         :return:
         """
+        src_dir = os.path.join(
+            self.staging_dir,
+            "steamapps",
+            "workshop",
+            "content",
+            str(self.arma3_app_id),
+            str(mod_id),
+        )
+        if not os.path.exists(src_dir):
+            raise Exception(
+                "Unable to locate mod: download failed or files deleted (check Celery logs)"
+            )
+
         self._delete_mod_(dst_dir)
         os.rename(
-            os.path.join(
-                self.staging_dir,
-                "steamapps",
-                "workshop",
-                "content",
-                str(self.arma3_app_id),
-                str(mod_id),
-            ),
+            src_dir,
             os.path.join(
                 self.dst_dir,
                 dst_dir,
@@ -873,8 +879,8 @@ class Arma3ServerHelper:
 class TaskHelper:
     def update_task_state(
         self,
-        task_type: int,
-        current_app: object,
+        task_type: str,
+        current_app,  # type: ignore
         level: str,
         schedule_id: int,
         msg: str,
@@ -885,6 +891,7 @@ class TaskHelper:
             The type of the task (for webhook notifications)
         :param current_app:
             The 'current_app' instance from Celery
+            Untyped because this is a local flask proxy which is not exposed and is a GIANT pain to pull in
         :param level:
             Logging level for the event
         :param schedule_id:
@@ -928,7 +935,7 @@ class TaskHelper:
         db.session.commit()
 
     @staticmethod
-    def send_webhooks(task_type, task_outcome) -> None:
+    def send_webhooks(task_type: str, task_outcome: str) -> None:
         notifications = []
         if task_type in ["server_restart", "server_start", "server_stop"]:
             notifications = Notification.query.filter(
